[VFX Attract] Fix crash on update.

+++ scripts/node_VFX_destroy/node_VFX_destroy.gml
@@ -7,9 +7,6 @@ function Node_VFX_Destroy(_x, _y, _group = noone) : Node_VFX_effector(_x, _y, _g
 	inputs[7].setVisible(false, false);
 	
 	function onAffect(part, str) {
-		var _sten = getInputData(5);
-		
-		if(random(1) < str * _sten)
-			part.kill();
+		if(random(1) < str * strength) part.kill();
 	}
 }
